+++ frontend/src/modules/builds/buildForm.js
@@ -1,6 +1,6 @@
 import { emptyInventorySlot } from '@/modules/builds/inventorySlots'
 
-export const equipmentUpgradeCount = 7
+export const equipmentUpgradeCount = 8
 
 export const weaponArcFields = Object.freeze([
   { fieldName: 'front_weapon_slots', labelKey: 'builds.create.weapons.front', altKey: 'builds.create.weapons.frontAlt' },
@@ -37,6 +37,7 @@ export function createBuildForm() {
     upgrade_5: '',
     upgrade_6: '',
     upgrade_7: '',
+    upgrade_8: '',
     lantern: '',
     research_upgrade_slot_unlocked: false,
     sailors: 0,